Rename `RejectionSubscriber` to `RejectionReactor`
... and fix test environment classes.
