+++ server/src/test/java/io/spine/server/rejection/given/CommandAwareReactor.java
@@ -20,29 +20,38 @@
 
 package io.spine.server.rejection.given;
 
+import com.google.protobuf.Empty;
 import io.spine.core.CommandContext;
+import io.spine.core.Commands;
+import io.spine.core.React;
 import io.spine.core.Rejection;
-import io.spine.core.Subscribe;
 import io.spine.test.rejection.ProjectRejections;
+import io.spine.test.rejection.command.RemoveOwner;
 
 import static io.spine.core.Rejections.getMessage;
 import static org.junit.Assert.assertEquals;
 
-public class ContextAwareSubscriber extends VerifiableSubscriber {
+public class CommandAwareReactor extends VerifiableReactor {
 
     private ProjectRejections.MissingOwner rejection;
+    private RemoveOwner command;
     private CommandContext context;
 
-    @Subscribe
-    public void on(ProjectRejections.MissingOwner rejection, CommandContext context) {
+    @React
+    public Empty on(ProjectRejections.MissingOwner rejection,
+                    RemoveOwner command,
+                    CommandContext context) {
         triggerCall();
         this.rejection = rejection;
+        this.command = command;
         this.context = context;
+        return Empty.getDefaultInstance();
     }
 
     @Override
     public void verifyGot(Rejection rejection) {
         assertEquals(getMessage(rejection), this.rejection);
+        assertEquals(Commands.getMessage(rejection.getContext().getCommand()), command);
         assertEquals(rejection.getContext().getCommand().getContext(), context);
     }
 }
